requirements.txt
pyyaml>=6.0
requests>=2.28.0
asyncio>=3.4.3
anthropic>=0.20.0
